sentinel/requirements.txt
ping3==4.0.4
dnspython==2.4.2
requests==2.31.0
paho-mqtt==1.6.1
pyyaml==6.0.1
schedule==1.2.1
speedtest-cli==2.1.3
